+++ Projects/UOContent/Mobiles/AI/BaseAI/BaseAI.cs
@@ -55,7 +55,7 @@ public enum ActionType
     Interact
 }
 
-public abstract class BaseAI
+public abstract partial class BaseAI
 {
     protected ActionType m_Action;
     public readonly BaseCreature m_Mobile;
@@ -131,64 +131,6 @@ public abstract class BaseAI
     {
         var name = m_Mobile.Name;
         return !string.IsNullOrEmpty(name) && speech.InsensitiveStartsWith(name);
-    }
-
-    public virtual void GetContextMenuEntries(Mobile from, ref PooledRefList<ContextMenuEntry> list)
-    {   // OSI has 16 tiles for opening the context menu.
-        if (!from.Alive || !m_Mobile.Controlled || !from.InRange(m_Mobile, 16))
-        {
-            return;
-        }
-    
-        if (from == m_Mobile.ControlMaster)
-        {
-            AddControlMasterEntries(ref list);
-        }
-        else if (m_Mobile.IsPetFriend(from))
-        {
-            AddPetFriendEntries(ref list);
-        }
-    }
-    
-    private void AddControlMasterEntries(ref PooledRefList<ContextMenuEntry> list)
-    {
-        var isDeadPet = m_Mobile.IsDeadPet; // Ordered list to match OSI.
-
-        // OSI has 14 tiles for context menu commands to be available.
-        // On OSI, commands grey out. Here, they do when out of range which is better, imo.
-        list.Add(new InternalEntry(3006111, 14, OrderType.Attack, !isDeadPet));     // Command: Kill
-        list.Add(new InternalEntry(3006108, 14, OrderType.Follow, true));           // Command: Follow
-        list.Add(new InternalEntry(3006107, 14, OrderType.Guard, !isDeadPet));      // Command: Guard
-        list.Add(new InternalEntry(3006112, 14, OrderType.Stop, true));             // Command: Stop
-        list.Add(new InternalEntry(3006114, 14, OrderType.Stay, true));             // Command: Stay
-
-        if (m_Mobile.CanDrop)
-        {
-            list.Add(new InternalEntry(3006109, 14, OrderType.Drop, !isDeadPet));   // Command: Drop
-        }
-
-        // This command is on OSI. However, the cliloc must be manually added to client files.
-        // simply returns a string giving instruction on how to rename pets.
-        list.Add(new InternalEntry(3006098, 14, OrderType.Rename, true));           // Rename
-    
-        if (!m_Mobile.Summoned && m_Mobile is not GrizzledMare)
-        {
-            list.Add(new InternalEntry(3006110, 14, OrderType.Friend, true));       // Add Friend
-            list.Add(new InternalEntry(3006099, 14, OrderType.Unfriend, true));     // Remove Friend
-            list.Add(new InternalEntry(3006113, 14, OrderType.Transfer, !isDeadPet)); // Transfer
-        }
-
-        // OSI has release command set to 13 tiles for no apparent reason.
-        list.Add(new InternalEntry(3006118, 13, OrderType.Release, true));          // Release
-    }
-    
-    private void AddPetFriendEntries(ref PooledRefList<ContextMenuEntry> list)
-    {
-        var isDeadPet = m_Mobile.IsDeadPet;
-    
-        list.Add(new InternalEntry(3006108, 14, OrderType.Follow, true));           // Command: Follow
-        list.Add(new InternalEntry(3006112, 14, OrderType.Stop, !isDeadPet));       // Command: Stop
-        list.Add(new InternalEntry(3006114, 14, OrderType.Stay, true));             // Command: Stay
     }
 
     public virtual void BeginPickTarget(Mobile from, OrderType order)
